chore: restore file naming scheme as per versio 1.x
