+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@acrolinx/sidebar-sdk",
   "private": false,
-  "version": "2.0.0-dev2",
+  "version": "2.0.0-dev3",
   "type": "module",
   "author": "Acrolinx",
   "url": "https://www.acrolinx.com",
