+++ src/adapters/AdapterInterface.ts
