+++ src/adapters/AutoBindAdapter.ts
@@ -22,8 +22,8 @@ import {
   ContentExtractionResult,
   ExtractContentForCheckOpts,
   SuccessfulCheckResult,
-} from './adapter-interface';
-import { MultiEditorAdapter, MultiEditorAdapterConfig } from './multi-editor-adapter';
+} from './AdapterInterface';
+import { MultiEditorAdapter, MultiEditorAdapterConfig } from './MultiEditorAdapter';
 
 // While making changes here make sure if you also need to do them in asynchronous version
 // of this adapter
