+++ src/adapters/CodeMirror5Adapter.ts
@@ -1,5 +1,6 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 /*
- * Copyright 2022-present Acrolinx GmbH
+ * Copyright 2017-present Acrolinx GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +15,7 @@
  * limitations under the License.
  */
 
+import { EditorFromTextArea } from 'codemirror';
 import { Check, DocumentSelection, Match, MatchWithReplacement } from '@acrolinx/sidebar-interface';
 import { lookupMatches } from '../lookup/diff-based';
 import { AlignedMatch } from '../utils/alignment';
@@ -23,32 +25,45 @@ import {
   ContentExtractionResult,
   ExtractContentForCheckOpts,
   SuccessfulCheckResult,
-} from './adapter-interface';
-import { EditorView } from '@codemirror/view';
+} from './AdapterInterface';
 import { encode } from 'entities';
 
-export type CodeMirror6AdapterConf = {
-  editor: EditorView;
+const FORMAT_BY_MIME_TYPE: { [mime: string]: string } = {
+  'text/html': 'HTML',
+  'text/xml': 'HTML',
+  'application/xml': 'XML',
+  'text/x-markdown': 'MARKDOWN',
+  'text/plain': 'TEXT',
+};
+
+const FORMAT_BY_MODE: { [mode: string]: string } = {
+  htmlmixed: 'HTML',
+  xml: 'XML',
+  markdown: 'MARKDOWN',
+};
+
+export type CodeMirrorAdapterConf = {
+  editor: CodeMirror.Editor | EditorFromTextArea;
   format?: string; // See CheckOptions.inputFormat
 };
 
-export class CodeMirror6Adapter implements AdapterInterface {
-  private config!: CodeMirror6AdapterConf;
+export class CodeMirror5Adapter implements AdapterInterface {
+  private config!: CodeMirrorAdapterConf;
   private currentContentChecking?: string;
   private lastContentChecked?: string;
   private formatDetectedByCheck: string | undefined;
 
-  constructor(conf: CodeMirror6AdapterConf) {
+  constructor(conf: CodeMirrorAdapterConf) {
     this.configure(conf);
   }
 
-  configure(partialConfig: Partial<CodeMirror6AdapterConf>) {
+  configure(partialConfig: Partial<CodeMirrorAdapterConf>) {
     const newConf = { ...this.config, ...partialConfig };
     this.validateConf(newConf);
     this.config = newConf;
   }
 
-  private validateConf(conf: CodeMirror6AdapterConf) {
+  private validateConf(conf: CodeMirrorAdapterConf) {
     if (!conf) {
       throw new Error('CodeMirrorAdapter config is missing');
     }
@@ -58,11 +73,18 @@ export class CodeMirror6Adapter implements AdapterInterface {
   }
 
   getContent() {
-    return this.config.editor.state.doc.toString();
+    return this.config.editor.getValue();
   }
 
   getFormat() {
-    return this.config.format || 'AUTO';
+    return this.config.format || this.getFormatFromCodeMirror() || 'AUTO';
+  }
+
+  private getFormatFromCodeMirror() {
+    return (
+      FORMAT_BY_MODE[this.config.editor.getDoc().getMode().name || ''] ||
+      FORMAT_BY_MIME_TYPE[this.config.editor.getOption('mode')! as string]
+    );
   }
 
   extractContentForCheck(opts: ExtractContentForCheckOpts): ContentExtractionResult {
@@ -75,7 +97,7 @@ export class CodeMirror6Adapter implements AdapterInterface {
 
   private getSelection(): DocumentSelection {
     return {
-      ranges: this.config.editor.state.selection.ranges.map(this.cmSelectionToRange),
+      ranges: this.getDoc().listSelections().map(this.cmSelectionToRange),
     };
   }
 
@@ -84,14 +106,11 @@ export class CodeMirror6Adapter implements AdapterInterface {
     this.lastContentChecked = this.currentContentChecking;
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  registerCheckCall(_checkInfo: Check) {
-    return;
-  }
+  registerCheckCall(_checkInfo: Check) {}
 
   private lookupMatchesOrThrow<T extends Match>(matches: T[]): AlignedMatch<T>[] {
     const alignedMatches = lookupMatches(this.lastContentChecked!, this.getContent(), matches, 'TEXT');
-    if (!alignedMatches || alignedMatches.length === 0) {
+    if (alignedMatches.length === 0) {
       throw Error('Selected flagged content is modified.');
     }
     return alignedMatches;
@@ -103,27 +122,19 @@ export class CodeMirror6Adapter implements AdapterInterface {
   }
 
   replaceRanges(_checkId: string, matchesWithReplacement: MatchWithReplacement[]) {
+    const doc = this.getDoc();
     const alignedMatches = this.lookupMatchesOrThrow(matchesWithReplacement);
     const escapeFunction = this.getEscapeFunction();
 
     let replacementLength = 0;
-    alignedMatches
-      .slice()
-      .reverse()
-      .forEach((match) => {
-        if (!isDangerousToReplace(this.lastContentChecked!, match.originalMatch)) {
-          const escapedReplacement = escapeFunction(match.originalMatch.replacement);
-          this.config.editor.dispatch({
-            changes: {
-              from: match.range[0],
-              to: match.range[1],
-              insert: escapedReplacement,
-            },
-          });
-
-          replacementLength += escapedReplacement.length;
-        }
-      });
+    for (const match of [...alignedMatches].reverse()) {
+      if (!isDangerousToReplace(this.lastContentChecked!, match.originalMatch)) {
+        const [start, end] = this.selectRange(match.range);
+        const escapedReplacement = escapeFunction(match.originalMatch.replacement);
+        doc.replaceRange(escapedReplacement, start, end);
+        replacementLength += escapedReplacement.length;
+      }
+    }
 
     this.selectRangeAndScroll([alignedMatches[0].range[0], alignedMatches[0].range[0] + replacementLength]);
   }
@@ -140,18 +151,31 @@ export class CodeMirror6Adapter implements AdapterInterface {
     }
   }
 
-  private selectRange(range: [number, number]): [number, number] {
-    this.config.editor.dispatch({ selection: { anchor: range[0], head: range[1] }, scrollIntoView: true });
-    return [range[0], range[1]];
+  private getDoc() {
+    return this.config.editor.getDoc();
+  }
+
+  private selectRange(range: [number, number]): [CodeMirror.Position, CodeMirror.Position] {
+    const doc = this.getDoc();
+    const startPos = doc.posFromIndex(range[0]);
+    const endPos = doc.posFromIndex(range[1]);
+    doc.setSelection(startPos, endPos);
+    return [startPos, endPos];
   }
 
   private selectRangeAndScroll(range: [number, number]) {
-    this.selectRange(range);
-    this.config.editor.focus();
+    const positionRange = this.selectRange(range);
+    const editor = this.config.editor;
+    editor.scrollIntoView(positionRange[0]);
+    editor.focus();
   }
 
-  private cmSelectionToRange = (selection: { from: number; to: number }): [number, number] => {
-    const range: [number, number] = [selection.from, selection.to];
+  private cmSelectionToRange = (selection: {
+    anchor: CodeMirror.Position;
+    head: CodeMirror.Position;
+  }): [number, number] => {
+    const doc = this.getDoc();
+    const range: [number, number] = [doc.indexFromPos(selection.anchor), doc.indexFromPos(selection.head)];
     range.sort((a, b) => a - b);
     return range;
   };
