+++ src/adapters/CodeMirror6Adapter.ts
@@ -1,6 +1,5 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 /*
- * Copyright 2017-present Acrolinx GmbH
+ * Copyright 2022-present Acrolinx GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +14,6 @@
  * limitations under the License.
  */
 
-import { EditorFromTextArea } from 'codemirror';
 import { Check, DocumentSelection, Match, MatchWithReplacement } from '@acrolinx/sidebar-interface';
 import { lookupMatches } from '../lookup/diff-based';
 import { AlignedMatch } from '../utils/alignment';
@@ -25,45 +23,32 @@ import {
   ContentExtractionResult,
   ExtractContentForCheckOpts,
   SuccessfulCheckResult,
-} from './adapter-interface';
+} from './AdapterInterface';
+import { EditorView } from '@codemirror/view';
 import { encode } from 'entities';
 
-const FORMAT_BY_MIME_TYPE: { [mime: string]: string } = {
-  'text/html': 'HTML',
-  'text/xml': 'HTML',
-  'application/xml': 'XML',
-  'text/x-markdown': 'MARKDOWN',
-  'text/plain': 'TEXT',
-};
-
-const FORMAT_BY_MODE: { [mode: string]: string } = {
-  htmlmixed: 'HTML',
-  xml: 'XML',
-  markdown: 'MARKDOWN',
-};
-
-export type CodeMirrorAdapterConf = {
-  editor: CodeMirror.Editor | EditorFromTextArea;
+export type CodeMirror6AdapterConf = {
+  editor: EditorView;
   format?: string; // See CheckOptions.inputFormat
 };
 
-export class CodeMirror5Adapter implements AdapterInterface {
-  private config!: CodeMirrorAdapterConf;
+export class CodeMirror6Adapter implements AdapterInterface {
+  private config!: CodeMirror6AdapterConf;
   private currentContentChecking?: string;
   private lastContentChecked?: string;
   private formatDetectedByCheck: string | undefined;
 
-  constructor(conf: CodeMirrorAdapterConf) {
+  constructor(conf: CodeMirror6AdapterConf) {
     this.configure(conf);
   }
 
-  configure(partialConfig: Partial<CodeMirrorAdapterConf>) {
+  configure(partialConfig: Partial<CodeMirror6AdapterConf>) {
     const newConf = { ...this.config, ...partialConfig };
     this.validateConf(newConf);
     this.config = newConf;
   }
 
-  private validateConf(conf: CodeMirrorAdapterConf) {
+  private validateConf(conf: CodeMirror6AdapterConf) {
     if (!conf) {
       throw new Error('CodeMirrorAdapter config is missing');
     }
@@ -73,18 +58,11 @@ export class CodeMirror5Adapter implements AdapterInterface {
   }
 
   getContent() {
-    return this.config.editor.getValue();
+    return this.config.editor.state.doc.toString();
   }
 
   getFormat() {
-    return this.config.format || this.getFormatFromCodeMirror() || 'AUTO';
-  }
-
-  private getFormatFromCodeMirror() {
-    return (
-      FORMAT_BY_MODE[this.config.editor.getDoc().getMode().name || ''] ||
-      FORMAT_BY_MIME_TYPE[this.config.editor.getOption('mode')! as string]
-    );
+    return this.config.format || 'AUTO';
   }
 
   extractContentForCheck(opts: ExtractContentForCheckOpts): ContentExtractionResult {
@@ -97,7 +75,7 @@ export class CodeMirror5Adapter implements AdapterInterface {
 
   private getSelection(): DocumentSelection {
     return {
-      ranges: this.getDoc().listSelections().map(this.cmSelectionToRange),
+      ranges: this.config.editor.state.selection.ranges.map(this.cmSelectionToRange),
     };
   }
 
@@ -106,11 +84,14 @@ export class CodeMirror5Adapter implements AdapterInterface {
     this.lastContentChecked = this.currentContentChecking;
   }
 
-  registerCheckCall(_checkInfo: Check) {}
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  registerCheckCall(_checkInfo: Check) {
+    return;
+  }
 
   private lookupMatchesOrThrow<T extends Match>(matches: T[]): AlignedMatch<T>[] {
     const alignedMatches = lookupMatches(this.lastContentChecked!, this.getContent(), matches, 'TEXT');
-    if (alignedMatches.length === 0) {
+    if (!alignedMatches || alignedMatches.length === 0) {
       throw Error('Selected flagged content is modified.');
     }
     return alignedMatches;
@@ -122,19 +103,27 @@ export class CodeMirror5Adapter implements AdapterInterface {
   }
 
   replaceRanges(_checkId: string, matchesWithReplacement: MatchWithReplacement[]) {
-    const doc = this.getDoc();
     const alignedMatches = this.lookupMatchesOrThrow(matchesWithReplacement);
     const escapeFunction = this.getEscapeFunction();
 
     let replacementLength = 0;
-    for (const match of [...alignedMatches].reverse()) {
-      if (!isDangerousToReplace(this.lastContentChecked!, match.originalMatch)) {
-        const [start, end] = this.selectRange(match.range);
-        const escapedReplacement = escapeFunction(match.originalMatch.replacement);
-        doc.replaceRange(escapedReplacement, start, end);
-        replacementLength += escapedReplacement.length;
-      }
-    }
+    alignedMatches
+      .slice()
+      .reverse()
+      .forEach((match) => {
+        if (!isDangerousToReplace(this.lastContentChecked!, match.originalMatch)) {
+          const escapedReplacement = escapeFunction(match.originalMatch.replacement);
+          this.config.editor.dispatch({
+            changes: {
+              from: match.range[0],
+              to: match.range[1],
+              insert: escapedReplacement,
+            },
+          });
+
+          replacementLength += escapedReplacement.length;
+        }
+      });
 
     this.selectRangeAndScroll([alignedMatches[0].range[0], alignedMatches[0].range[0] + replacementLength]);
   }
@@ -151,31 +140,18 @@ export class CodeMirror5Adapter implements AdapterInterface {
     }
   }
 
-  private getDoc() {
-    return this.config.editor.getDoc();
-  }
-
-  private selectRange(range: [number, number]): [CodeMirror.Position, CodeMirror.Position] {
-    const doc = this.getDoc();
-    const startPos = doc.posFromIndex(range[0]);
-    const endPos = doc.posFromIndex(range[1]);
-    doc.setSelection(startPos, endPos);
-    return [startPos, endPos];
+  private selectRange(range: [number, number]): [number, number] {
+    this.config.editor.dispatch({ selection: { anchor: range[0], head: range[1] }, scrollIntoView: true });
+    return [range[0], range[1]];
   }
 
   private selectRangeAndScroll(range: [number, number]) {
-    const positionRange = this.selectRange(range);
-    const editor = this.config.editor;
-    editor.scrollIntoView(positionRange[0]);
-    editor.focus();
+    this.selectRange(range);
+    this.config.editor.focus();
   }
 
-  private cmSelectionToRange = (selection: {
-    anchor: CodeMirror.Position;
-    head: CodeMirror.Position;
-  }): [number, number] => {
-    const doc = this.getDoc();
-    const range: [number, number] = [doc.indexFromPos(selection.anchor), doc.indexFromPos(selection.head)];
+  private cmSelectionToRange = (selection: { from: number; to: number }): [number, number] => {
+    const range: [number, number] = [selection.from, selection.to];
     range.sort((a, b) => a - b);
     return range;
   };
