+++ src/index.ts
@@ -19,52 +19,52 @@
 export { AcrolinxPlugin } from './acrolinx-plugin';
 export type { AcrolinxPluginConfig } from './acrolinx-plugin-config';
 export { autoBindFloatingSidebar, autoBindFloatingSidebarAsync } from './autobind-plugin';
-export type { AdapterInterface, AsyncAdapterInterface, CommonAdapterInterface } from './adapters/adapter-interface';
-export { InputAdapter } from './adapters/input-adapter';
-export { ContentEditableAdapter } from './adapters/content-editable-adapter';
-export { AsyncContentEditableAdapter as StateBasedContentEditableAdapter } from './adapters/async-content-editable-adapter';
-export { AbstractRichtextEditorAdapter } from './adapters/abstract-rich-text-editor-adapter';
-export { CKEditor5Adapter } from './adapters/ckeditor-5-adapter';
-export { CKEditor4Adapter as CKEditorAdapter } from './adapters/ckeditor-4-adapter';
-export { TinyMCEAdapter } from './adapters/tinymce-adapter';
-export { AutoBindAdapter } from './adapters/autobind-adapter';
-export { AsyncAutoBindAdapter } from './adapters/async-autobind-adapter';
-export { MultiEditorAdapter } from './adapters/multi-editor-adapter';
-export { AsyncMultiEditorAdapter } from './adapters/async-multi-editor-adapter';
+export type { AdapterInterface, AsyncAdapterInterface, CommonAdapterInterface } from './adapters/AdapterInterface';
+export { InputAdapter } from './adapters/InputAdapter';
+export { ContentEditableAdapter } from './adapters/ContentEditableAdapter';
+export { AsyncContentEditableAdapter as StateBasedContentEditableAdapter } from './adapters/AsyncContentEditableAdapter';
+export { AbstractRichtextEditorAdapter } from './adapters/AbstractRichtextEditorAdapter';
+export { CKEditor5Adapter } from './adapters/CKEditor5Adapter';
+export { CKEditor4Adapter as CKEditorAdapter } from './adapters/CKEditor4Adapter';
+export { TinyMCEAdapter } from './adapters/TinyMCEAdapter';
+export { AutoBindAdapter } from './adapters/AutoBindAdapter';
+export { AsyncAutoBindAdapter } from './adapters/AsyncAutoBindAdapter';
+export { MultiEditorAdapter } from './adapters/MultiEditorAdapter';
+export { AsyncMultiEditorAdapter } from './adapters/AsyncMultiEditorAdapter';
 export { createPluginMessageAdapter } from './message-adapter/message-adapter';
 export { loadSidebarCode } from './utils/sidebar-loader';
 export { getSelectionHtmlRanges } from './utils/check-selection';
 export { lookupMatches } from './lookup/diff-based';
 export { extractTextDomMapping } from './utils/text-dom-mapping';
-export { CodeMirror6Adapter } from './adapters/codemirror-6-adapter';
-export { CodeMirror5Adapter as CodeMirrorAdapter } from './adapters/codemirror-5-adapter';
+export { CodeMirror6Adapter } from './adapters/CodeMirror6Adapter';
+export { CodeMirror5Adapter as CodeMirrorAdapter } from './adapters/CodeMirror5Adapter';
 export { initFloatingSidebar } from './floating-sidebar/floating-sidebar';
 export { AsyncLocalStorage } from './floating-sidebar/async-storage';
-export { XeditorAdapter } from './adapters/x-editor-adapter';
+export { XeditorAdapter } from './adapters/XeditorAdapter';
 
 import { AcrolinxPlugin } from './acrolinx-plugin';
 import { autoBindFloatingSidebar, autoBindFloatingSidebarAsync } from './autobind-plugin';
-import { InputAdapter } from './adapters/input-adapter';
-import { ContentEditableAdapter } from './adapters/content-editable-adapter';
-import { AsyncContentEditableAdapter } from './adapters/async-content-editable-adapter';
-import { AbstractRichtextEditorAdapter } from './adapters/abstract-rich-text-editor-adapter';
-import { CKEditor5Adapter } from './adapters/ckeditor-5-adapter';
-import { CKEditor4Adapter } from './adapters/ckeditor-4-adapter';
-import { TinyMCEAdapter } from './adapters/tinymce-adapter';
-import { AutoBindAdapter } from './adapters/autobind-adapter';
-import { AsyncAutoBindAdapter } from './adapters/async-autobind-adapter';
-import { MultiEditorAdapter } from './adapters/multi-editor-adapter';
-import { AsyncMultiEditorAdapter } from './adapters/async-multi-editor-adapter';
+import { InputAdapter } from './adapters/InputAdapter';
+import { ContentEditableAdapter } from './adapters/ContentEditableAdapter';
+import { AsyncContentEditableAdapter } from './adapters/AsyncContentEditableAdapter';
+import { AbstractRichtextEditorAdapter } from './adapters/AbstractRichtextEditorAdapter';
+import { CKEditor5Adapter } from './adapters/CKEditor5Adapter';
+import { CKEditor4Adapter } from './adapters/CKEditor4Adapter';
+import { TinyMCEAdapter } from './adapters/TinyMCEAdapter';
+import { AutoBindAdapter } from './adapters/AutoBindAdapter';
+import { AsyncAutoBindAdapter } from './adapters/AsyncAutoBindAdapter';
+import { MultiEditorAdapter } from './adapters/MultiEditorAdapter';
+import { AsyncMultiEditorAdapter } from './adapters/AsyncMultiEditorAdapter';
 import { createPluginMessageAdapter } from './message-adapter/message-adapter';
 import { loadSidebarCode } from './utils/sidebar-loader';
 import { getSelectionHtmlRanges } from './utils/check-selection';
 import { lookupMatches } from './lookup/diff-based';
 import { extractTextDomMapping } from './utils/text-dom-mapping';
-import { CodeMirror6Adapter } from './adapters/codemirror-6-adapter';
-import { CodeMirror5Adapter } from './adapters/codemirror-5-adapter';
+import { CodeMirror6Adapter } from './adapters/CodeMirror6Adapter';
+import { CodeMirror5Adapter } from './adapters/CodeMirror5Adapter';
 import { initFloatingSidebar } from './floating-sidebar/floating-sidebar';
 import { AsyncLocalStorage } from './floating-sidebar/async-storage';
-import { XeditorAdapter } from './adapters/x-editor-adapter';
+import { XeditorAdapter } from './adapters/XeditorAdapter';
 
 export interface AcrolinxSidebarIntegration {
   AcrolinxPlugin: typeof AcrolinxPlugin;
