+++ tests/browser/adapter-setups/adapter-setup.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { AdapterInterface } from '../../../src/adapters/adapter-interface';
+import { AdapterInterface } from '../../../src/adapters/AdapterInterface';
 
 export type DoneCallback = () => void;
