+++ tests/browser/ckeditor-4-adapter.test.ts
@@ -15,10 +15,10 @@
  */
 
 import { expect, describe, beforeEach, afterEach, it } from 'vitest';
-import { AdapterInterface } from '../../src/adapters/adapter-interface';
+import { AdapterInterface } from '../../src/adapters/AdapterInterface';
 import { assertEditorText, givenATextAync, htmlStringToElements } from './utils/util';
 import { isChrome } from '../../src/utils/detect-browser';
-import { AbstractRichtextEditorAdapter } from '../../src/adapters/abstract-rich-text-editor-adapter';
+import { AbstractRichtextEditorAdapter } from '../../src/adapters/AbstractRichtextEditorAdapter';
 import { containsEmptyTextNodes, getMatchesWithReplacement } from './utils/test-utils';
 import { MatchWithReplacement } from '@acrolinx/sidebar-interface';
 import { dummyCheckId, NON_BREAKING_SPACE } from './adapter-setups/constants';
